LU-680 instance <-> mount point mapping from lfs
A new option to 'lfs' has been created to return the mapping
between Lustre filesystem instance and mount points. The option
is 'mount' and is called without arguements: 'lfs mount'

All the Lustre mounts are iterated over. For each mount point an
existing function is called to provide the mapping:
llapi_file_get_lov_uuid. The result from this call is tidied up
with an existing and new helper function and printed to the screen.

Documentation for 'lfs mount' is included inline.

Change-Id: Iab8ff12d604c7ce853f3c204b455e3b641f659f4
Signed-off-by: Richard Henwood <rhenwood@whamcloud.com>