LU-680 instance <-> mount point mapping from lfs
A new option to 'lfs' has been created to return the mapping
between Lustre filesystem instance and paths. The option
is 'getname' and it may be called with or without arguements.

'lfs getname' without arguments returns the instances of all
Lustre mount points.

'lfs getname [path...]' returns the instance of each specified
path. If the path is not a Lustre instance 'No such device' is
returned.

For each path, an existing function is called to provide
the mapping: llapi_file_get_lov_uuid. The result from this call
is tidied up and printed to the screen.

Documentation for 'lfs getname' is included inline.

Signed-off-by: Richard Henwood <rhenwood@whamcloud.com>
Change-Id: Iab8ff12d604c7ce853f3c204b455e3b641f659f4