LU-682 test for tar patched for completely sparse files.
test 102l: FAIL if tar does not handle sparse files with
zero blocks in a optimized way.

The test works by creating a large file with zero blocks.
This file is then tar'ed using 'tar' from the path. If
a un-patched version of tar is called, the tar command
takes >0.01 seconds provided the local disk system is
slower than 400GB/sec. If a patched tar is present tar
completes in <0.01 sec.

Change-Id: I48a93324aaca400d7006f52ac6b80ead0fe23c19
Signed-off-by: Richard Henwood <rhenwood@whamcloud.com>