LU-723 Move WALK_SPACE_HAS_DATA_SEM into ldiskfs
In order for Lustre to correctly build its ldiskfs lvfs code, it needs
some insight into whether the ext4_ext_walk_space function which ldiskfs
was build against downs the i_data_sem semaphore internally.

Rather than relying on an ugly autoconf test to probe the internals of
the ext4_ext_walk_space function and then set an AC_DEFINE, the
responsibility of defining the correct macro was moved into the ldiskfs
stack. This means that rather than checking, we make a perfectly
reasonable assumption regarding the need of the macro on 2.6.32 and
newer based kernels.

The main reason this change is needed is to facilitate building Lustre
without full kernel sources using an external ldiskfs development
package.

Signed-off-by: Prakash Surya <surya1@llnl.gov>
Change-Id: I9bd12184671955289fd8b93117fef09a8f644af0