ORI-350 Move lmv/mdc/osc/lov quota code out of lquota
All quota code was initially put into a separate kernel module since quota
might be released under a different license. That's not relevant any more and
we can now move the client-side quota code to the regular lustre modules.
This removes useless indirections and makes the code easier to read and to
maintain.

Change-Id: If898a46db9158edb8e4eaf855f1ed98db97330f0
Signed-off-by: Niu Yawei <niu@whamcloud.com>
Signed-off-by: Johann Lombardi <johann@whamcloud.com>