LU-728 pdsh HOSTLIST test suite support
For large systems it becomes tedious to list all the nodes or devices. The
application pdsh uses a HOSTLIST form to make listing nodes more compact. This
patch takes data in the HOSTLIST format and creates a expanded list for the
shell scripts in the test suite to use. Also this gives the option for the
pdsh commands to use the nodes list in HOSTLIST format directly. Its is also
possible to use the HOSTLIST format for devices as well as nodes.

Signed-off-by: James Simmons <uja.ornl@gmail.com>
Change-Id: Ia7a6c95ab4e701a17fc2b80d8443a5ff79da3f3c