LU-769 ptlrpc: conf-sanity test_23a failure
The problem is clear: in original implementation, if a signal is already
existent, it will go for uninterruptible version of LWI. This causes the
waiting process will never be interrupted.

I fixed this problem by removing check of cfs_signal_pending(), but it
seems this piece of code needs more fixing.

Change-Id: I28e9dc1969f1ab03800d1287392ab04459e8e1d5
Signed-off-by: Jinshan Xiong <jay@whamcloud.com>