LU-483: get_ec_devno returns trash
Components: utils/obd.c ldiskfs/configure.ac obdfilter-survey

plus a few cleanups:

lustre/utils/obd.c:
shmem_bump() modifies bumped_running without lock

Each of several threads could sleep after the shmem_unlock call
allowing following threads to see bumped_running set to false. It
should be set to true inside the guarded "then" clause still under the
shmem_lock() protection. A side effect is that it will only ever be
set to true once.

====
lustre/autoconf/lustre-core.m4:
ldiskfs_is_ext4 is unused

"--with-ext4" is an obsolete option.  This is left over from that.

====
.gitignore:
* Ignore files created by "patch" program.

====
lustre-iokit/obdfilter-survey/obdfilter-survey:
Redirecting stdin in a while loop can trigger subshell.

That causes variables changed in the loop to loose their changes.
(POSIX does not require "while" loops in shell scripts to be performed
in the same environment when stdin is redirected.  "bash" does not
commit to maintaining current behavior.)  Therefore, read input from
fd 3.  Also:

* Eliminate unnecessary fork/exec calls since the analysis
  can readily be done in the same shell process.
* Ensure that all error paths do emit ERROR to stdout.
* Ensure "libecho" comes from obdfilter-survey dir.

====
lustre-iokit/obdfilter-survey/libecho:
get_ec_devno() in libecho mishandles results

The results from this function are returned by echoing them to stdout. This
needs to be done more safely because other functions that it invokes
(e.g. "remote_shell") can drop detritus into stdout (viz., a terminal control
escape sequence) that mutilates results.

The fundamental bug is that readline(3GNU) should *never* emit terminal
control sequences to stdout.  It should be stderr, especially when stdin is
not a terminal.  We won't be fixing the readline function, but we can mitigate
the bug by dumping unneeded stdout output to stderr or /dev/null.

So, the solution is to clone stdout fd, redirect stdout to an alternate fd
(stderr) and echo results to the cloned fd.

Since this works only as a code fragment with redirected stdout, it must,
perforce, be running in a subshell. The "return" statements need to be "exit"
statements, with the proper success/fail indication. This is not currently
used, but should be more reliable than checking the result count.

====
Reviewed-by: Alexey Lyashkov <alexey_lyashkov@xyratex.com>
Signed-off-by: Bruce Korb <bruce_korb@xyratex.com>
Change-Id: I50137c9ded816aac792aba4a0a79f59c28d16e01