LU-876 test: several improvement to ost-pool.sh
Fix a bug in the output parsing of createmany in test 18 causing
crazy results like this:
Avg time taken for 30000 creates without pool: 1322089193.67
Avg time taken for 30000 creates with pool: 1322089359.43
Avg time taken for 30000 creates with missing pool: 1322089580.10

createmany actually prints several lines when creating more than 10k
files, causing this bug.
Test 18 has also been changed to create a less files when SLOW=no
since this test currently take ~30min to complete with the default
parameters.
Test 14 is another one taking a lot of time and it is because it
fills up OST0 with dd, so the usual size check to skip this test has
been added.

Signed-off-by: Johann Lombardi <johann@whamcloud.com>
Change-Id: I887c3eb86b2e6103aef62f4d059c6fe3d355e719