LU-708 utils: refactor mkfs/tunefs/mount
Port of ORI-294. Original patch locates at
http://review.whamcloud.com/1357

Refactor the mkfs.lustre, tunefs.lustre, mount.lustre utilities
to cleanly support multiple backend filesystems.  While this
commit initially appears large it does not introduce any
functional changes.  It does however do the following.

* Global variables consolidated in mkfs_opts/mount_opts
  structures so this information may be cleanly passed to
  common functions provided by mount_util.c.

* Functions which were defined once in each utility have
  been made common and consolidated in mount_utils.c.  In
  a few cases the common function was made mode generic.

* Split mkfs.lustre and tunefs.lustre in seperate .c files
  for clarity.  This removes the need for the TUNEFS #define.

* General cleanup.  This includes removing unused variables
  and function, renaming functions for consistently, and
  removing unused headers.

The bulk of this ldiskfs specific code was simply moved from the
existing utilties and consolidated in mount_utils_ldiskfs.c.
Minimal changes were made to support the refactored infrastructure.

A small fix was also made to the lustre spec file.  It was
mistakenly checking if the libiam.c file exists and then assuming
it was safe to add the libiam.a library.  However, the library
will only be build when the the ldiskfs osd code is enabled so
it should check for the .a file.

Author: Brian Behlendorf <behlendorf1@llnl.gov>

test-framework should not resolve the path to utils like mkfs.lustre
and tunefs.lustre on a lustre client since client-only RPMs no longer
have those tools installed (check ORI-415 for more background).

Author: Johann Lombardi <johann@whamcloud.com>

Signed-off-by: Bobi Jam <bobijam@whamcloud.com>
Change-Id: Ic98c18f3072cf4a3d6a17c10763f8a70ba6dc5c7