LU-915 tests: MKFSOPTS fix for test suite
While testing LU-80 I had to provide my own MKFSOPT string
for the mkfsoptions in local.sh. What was discovered was
the order is incorrect in local.sh which caused my formats
to fail. The reason is my extra MKFSOPT never ended up in
the mkfsoptions string but was appended directly to the
MKFSOPT string instead. As a side improvement I also make
handling device-size easier so all you have to do is pound
out "OBD"SIZE.

Signed-off-by: James Simmons <uja.ornl@gmail.com>
Change-Id: If30d2a692431dcdb3631fc9c1e77da5ff2792881