LU-169 layout: add generation number to LOV EA
This patch shrinks lov_mds_md_v*::lmm_stripe_count to 16 bits and use
the remaining 16 bits to store a generation number for the layout.
This generation will be used in conjunction the layout lock to allow
clients to detect when the file layout has changed.
The layout generation starts at 0 and will be bumped each time the
file layout is altered.
For backward compatibility, the generation number is reset to 0 if the
client does not support OBD_CONNECT_LAYOUTLOCK.
Similarly, the generation is set to -1 for directory using a stripe
count of -1.

This patch also adds an incompat flag for lustre version which cannot
deal with a 16-bit lmm_stripe_count. This flag will be set in
last_rcvd only once we start modifying file layouts, but at least the
current version should be confused by a 16-bit stripe count and a
non-null generation.

Signed-off-by: Jacques-Charles Lafoucriere <jacques-charles.lafoucriere@cea.fr>
Change-Id: Ic40083227057eba565287d1a10890875b8a96c13
Signed-off-by: Johann Lombardi <johann@whamcloud.com>