LU-80 utils: Large stripe support
Currently a file can be stripped across OSTs up to a limit of
160 stripes due to the ldiskfs xattr size limit of 4096 bytes.
This limit will be increased to 2000 stripes or more by increasing
the maximum xattr size for the MDT.

During testing, issues emerged with clients that are not expecting
more than 160 stripes. This patch allows clients to interoperate with
servers that do not support large xattrs, and also servers with
support with at least 2000 stripes, though it is intended not to have
any hard upper limit.

Change-Id: Idbaeb98919aea6b4cd375b881ed87661034d9394
Signed-off-by: James Simmons <uja.ornl@gmail.com>