ORI-448 tests: replay_barrier: Alloc seq on MOUNT2
In order to avoid sequence file updates after a target is made
read-only, replay_barrier() creates a file in "MOUNT" on every client
node.  This is not enough because the "MOUNT2" clients may also trigger
sequence file updates.  This patch makes sure the trick is applied to
"MOUNT2" clients as well.

Change-Id: Ib81af9462deb310d11332239ebc2cdad5463d3b2
Signed-off-by: Li Wei <liwei@whamcloud.com>