LU-904 ptlrpc: redo io on -EINPROGRESS
When server return -EINPROGRESS for a write RPC, the client should keep
resending the RPC until server return other error code or the client is
evicted.

This is required by the new quota design: when a write on OST can't
acquire quota from master for broken network, it should return
-EINPROGRESS to inform the client to retry write infinitely.

This patch also fixed the defect of redo io RPC can't be aborted during
eviction, in a lightweight manner.

Signed-off-by: Niu Yawei <niu@whamcloud.com>
Change-Id: Iea393cb1ea55e9d006f52dbfc39a2b9a3670d682