LU-577 tests: FAIL replay-single test_70b rundbench load
Test 70b for replay-single assumes that lustre is mounted on
/mnt/lustre which is not the case for us. This patch passes
the proper MOUNT. The test also was not using the standard
DIR/tdir setup which had generated data files not being
cleaned up. Increased the sleep period to match dbench's
warm up period. This gives dbench a change to start up when
using many clients.

Signed-off-by: James Simmons <uja.ornl@gmail.com>
Change-Id: Iefdfe4bf47d322e782e138cf6ae17240b3004d8b