LU-169 lov: add lsm refcounting
This patch adds a reference counter to the lsm structure.
Each time a lsm is used, a reference must be taken with lsm_get/addref
(lsm_addref() can be used when we already own a reference on the lsm).
This reference can be released via lsm_put/decref. The lsm is freed
when the last reference is dropped.

The layout still cannot be changed since the layout lock is not
enqueued by client nodes yet, but a proper tracking of lsm users is a
requirement to support migration (or any other features involving a
layout change) in the future.

Signed-off-by: Jacques-Charles Lafoucriere <jacques-charles.lafoucriere@cea.fr>
Signed-off-by: Johann Lombardi <johann@whamcloud.com>
Signed-off-by: Liang Zhen <liang@whamcloud.com>
Change-Id: I156b4cc2dc82bb15ae8107cf9842f100048c03d4