LU-1028 tests: re-enable sanityn.sh test_16 (fsx)
The sanityn.sh version of fsx in test_16 provides useful test
coverage because it is running an IO load that is doing unaligned
file access from multiple clients (mountpoints) and also verifying
data correctness.  This has exposed several bugs in the orion branch,
and should also be enabled by default for all landings on master.

A longer version of the test is run in sanity-benchmark.sh for
better test coverage.  Check that fsx completes successfully,
otherwise it may fail without reporting an error to the test
framework.

Signed-off-by: Andreas Dilger <adilger@whamcloud.com>
Change-Id: I26dd91cacc2596f6e84f4c32362befa1e339cab0