LU-1032 build: Add dkms support to debian client packages.
This change adds dkms support to the debian/ubuntu packaging
infrastructure.

dkms is a cross distro mechanism for the building and maintenance
of out-of-tree kernel modules. (http://linux.dell.com/dkms/)

The current debian/ubuntu kernel module package uses
module-assistant, which produces a deb which is tied to a specific
kernel version, requiring the user to manually rebuild packages
when the kernel is changed.

The dkms package will automatically build kernel modules for all
installed kernels. The package also contains triggers so that when
a new kernel is installed, lustre client modules are automatically
build for the new kernel. This reduces maintenance overhead on
client machines.

Signed-off-by: Guy Coates <gmpc@sanger.ac.uk>
Change-Id: I6b576881d2518fb2765b41b3a336e464f0101932