LU-459 tests: quiet spurious console/test messages
Quiet a number of frequent and useless messages printed to the test
output and/or the console.  These distract from understanding what
is happening in the test and provide little, if any, value.

The "acl" mount option on the client has been deprecated since 1.8.0,
since it was added in mid-1.6, and is driven by the MDS config only.

Use textual debug/subsystem flags instead of numeric flags.  This has
been possible through all of 1.8, but is deprecated in 2.x so that
we can move to a more sophisticated debug mechanism in the future.

Signed-off-by: Andreas Dilger <adilger@whamcloud.com>
Change-Id: I81bed13c22465dc179e19f5a6714b732d68c500c