LU-1050 o2iblnd: fix checking order of rdma_create_id() argument
Need to change the checking order for number of argument of
rdma_create_id(). current ordering in lustre-lnet.m4
1) checking supported arguments of rdma_create_id()
2) detection of O2IBPATHS if o2ib is enabled
3) check rdma_create_id() works or not

1)checking rdma_create_id() arguments always compiled with
in-kernel-tree OFED header. If the number of arguments is four
with rhel6.2 kernel, this only works if O2IBPATHS pointed to
in-kernel-tree OFED.
If O2IBPATHS is pointed to OFED-1.5.4, it failed due to external
OFED still only supports three arguments in rdma_create_id().
Checking supported arguments of rdma_create_id() should run after
2)detection of O2IBPATHS to find correct number of arguments.

Change-Id: I0782183f15f58647291518a4222610601083c369
Signed-off-by: Shuichi Ihara <sihara@ddn.com>