TT-282 build: add Test-Parameters tag for autotest
Allow Test-Parameters line to specify extra tests that should be run
to properly validate a change made to code or a regression test.

The Test-Parameters line can be quite long, depending on the number
of parameters specified.  A single set of test parameters can be
split over multiple lines by escaping the linefeed, though the
individual lines should still be kept to a reasonable length.

Signed-off-by: Andreas Dilger <adilger@whamcloud.com>
Change-Id: I65e1d41ad0f487df3753f25d9111ade9a0af08b1