LU-1110 fid: add full support for open-by-fid
Open-by-fid parent 'fid' is a in-memory struct, but to mimic a real
in stable storage object to be able to op-by-fid, it's marked
LOHA_EXISTS to pass many checks in operations handling, however this
may cause dt operations. Add a flag LOHA_FAKE to define such object
more clearly, and forbid link/unlink/rename/create operations if
parent object is such kind of object.

Add more sanity tests for op-by-fid.

Signed-off-by: Lai Siyao <laisiyao@whamcloud.com>
Change-Id: I10869923e861ac0165886a48880f58e1269554ec