LU-1206: mdt: Fix error handling in mdt_mfd_open
In mdt_mfd_open if the mo_open() call failed or we could not allocate
mfd, we also need to undo write/exec reference count in order to
not mess up with subsequent exec/write accesses.
While we are at it, I also moved io_epoch opening (for som)
to the very end of the function as I don't see a very easy way to undo it
should the failure occur.

Signed-off-by: Prakash Surya <surya1@llnl.gov>
Signed-off-by: Oleg Drokin <green@whamcloud.com>
Change-Id: I3bd98bd68368b48f2afaa7bb450d3a9947c992ac